zap: EditList로 통합 및 가져오는 로직 수정
