+++ src/components/list/ListCard.tsx
@@ -1,86 +1,114 @@
-import { Check, Heart, MessageSquare } from 'lucide-react';
+import { Heart, MessageSquare } from 'lucide-react';
 import Avartar from '../ui/Avartar';
-import type { PostType } from '../../types';
-import { format } from 'date-fns';
+import type { PostType, User } from '../../types';
 import { previewMarkdown } from '../../utils/markdown';
+import dateFormat from '../../utils/dateFormat';
+import { useEffect, useState } from 'react';
+import { useModalStore } from '../../stores/modalStore';
+import { useNavigate } from 'react-router';
+import { useAuthStore } from '../../stores/authStore';
+import { getUser } from '../../api/userApi';
 
-// 임시로 지정한 props 입니다
-export default function ListCard({
-  data,
-  solve, //문제 풀이여부 확인
-}: {
-  data: PostType;
-  solve?: boolean;
-}) {
+export default function ListCard({ data }: { data: PostType }) {
+  const session = useAuthStore((state) => state.session);
+  const [isPending, setPending] = useState(false);
+  const [me, setMe] = useState<User>(null);
+
+  const { setLogInModal } = useModalStore();
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        setPending(true);
+        const myData = await getUser(session?.user.id as string);
+        setMe(myData);
+      } catch (e) {
+        console.error(e);
+      } finally {
+        setPending(false);
+      }
+    };
+    fetchData();
+  }, [data, session]);
+
+  const handleClick = () => {
+    if (!session?.user.id) {
+      setLogInModal(true);
+      return;
+    }
+    console.log(me);
+    navigate(`/questions/${data.id}`);
+  };
   return (
     <>
-      {/* 스켈레톤 */}
-      {/* <div className="w-full rounded-sm border border-[#ccc]">
-        <div className="px-3 pt-3.5 pb-3 md:px-4 md:pt-4 md:pb-3.5">
-          <div className="mb-2.5 h-3.5 w-2/3 bg-gray-200 md:h-4.5 lg:h-5.5"></div>
-          <div className="mb-1 h-3 w-full bg-gray-200 md:h-4 lg:h-5"></div>
-          <div className="mb-2.5 h-3 w-4/5 bg-gray-200 md:h-4 lg:h-5"></div>
-          <div className="h-2.5 w-1/3 bg-gray-200 md:h-3.5 lg:h-4.5"></div>
-        </div>
-        <div className="flex w-full items-center gap-2 border-t border-[#ccc] px-3 py-2 md:px-4 md:py-2.5">
-          <div className="h-[24px] w-[24px] rounded-full bg-gray-200 md:h-[26px] md:w-[26px] lg:h-[28px] lg:w-[28px]"></div>
-          <div className="h-3 w-1/2 bg-gray-200 md:h-4 lg:h-5"></div>
-        </div>
-      </div> */}
-
-      <div className="w-full rounded-sm border border-[#ccc]">
-        <div className="px-3 pt-3.5 pb-3 md:px-4 md:pt-4 md:pb-3.5">
-          <div className="flex gap-2.5">
-            <div className="w-[calc(100%-110px)] md:w-[calc(100%-130px)] lg:w-[calc(100%-150px)]">
-              <p className="mb-2.5 text-sm font-semibold md:text-base lg:text-lg">
-                {data.title}
-              </p>
-              <p className="mb-2.5 line-clamp-2 text-xs md:text-sm lg:text-base">
-                {data.content
-                  ? previewMarkdown(data.content).slice(0, 100)
-                  : ''}
-              </p>
-            </div>
-            {data.image && (
-              <div className="ml-auto h-[65px] w-[85px] shrink-0 overflow-hidden md:h-[75px] md:w-[105px] lg:h-[85px] lg:w-[125px]">
-                <img src={data.image} className="h-full w-full object-cover" />
-              </div>
-            )}
+      {isPending ? (
+        <div className="w-full rounded-sm border border-[#ccc]">
+          <div className="px-3 pt-3.5 pb-3 md:px-4 md:pt-4 md:pb-3.5">
+            <div className="mb-2.5 h-3.5 w-2/3 bg-gray-200 md:h-4.5 lg:h-5.5"></div>
+            <div className="mb-1 h-3 w-full bg-gray-200 md:h-4 lg:h-5"></div>
+            <div className="mb-2.5 h-3 w-4/5 bg-gray-200 md:h-4 lg:h-5"></div>
+            <div className="h-2.5 w-1/3 bg-gray-200 md:h-3.5 lg:h-4.5"></div>
           </div>
-          <ul className="mb-2.5 flex gap-3">
-            <li className="rounded-sm bg-[var(--color-gray1)] px-2 py-0.5 text-[10px] text-[var(--color-gray4)] md:text-xs lg:text-sm">
-              태그명
-            </li>
-            <li className="rounded-sm bg-[var(--color-gray1)] px-2 py-0.5 text-[10px] text-[var(--color-gray4)] md:text-xs lg:text-sm">
-              태그명
-            </li>
-          </ul>
-          <div className="flex items-end">
-            <span className="text-[10px] text-[var(--color-gray3)] md:text-xs lg:text-sm">
-              {format(new Date(data.created_at), 'yyyy-MM-dd')}
-            </span>
-            <div className="ml-auto flex shrink-0 gap-3">
-              <div className="flex items-center gap-1">
-                <Heart className="w-3.5 md:w-4 lg:w-4.5" />
-                <span className="text-[10px] md:text-xs lg:text-sm">5</span>
-              </div>
-              <div className="flex items-center gap-1">
-                <MessageSquare className="w-3.5 md:w-4 lg:w-4.5" />
-                <span className="text-[10px] md:text-xs lg:text-sm">5</span>
-              </div>
-            </div>
+          <div className="flex w-full items-center gap-2 border-t border-[#ccc] px-3 py-2 md:px-4 md:py-2.5">
+            <div className="h-[24px] w-[24px] rounded-full bg-gray-200 md:h-[26px] md:w-[26px] lg:h-[28px] lg:w-[28px]"></div>
+            <div className="h-3 w-1/2 bg-gray-200 md:h-4 lg:h-5"></div>
           </div>
         </div>
-        <div className="flex w-full items-center border-t border-[#ccc] px-3 py-2 md:px-4 md:py-2.5">
-          <Avartar />
-          {solve && (
-            <p className="ml-auto flex items-center gap-1 text-[10px] md:text-xs lg:text-sm">
-              <Check className="w-4 text-[var(--color-green-info)] md:w-5 lg:w-6" />
-              풀이됨
-            </p>
-          )}
+      ) : (
+        <div onClick={handleClick} className="cursor-pointer">
+          <div className="w-full rounded-sm border border-[#ccc]">
+            <div className="px-3 pt-3.5 pb-3 md:px-4 md:pt-4 md:pb-3.5">
+              <div className="flex gap-2.5">
+                <div className="w-[calc(100%-110px)] md:w-[calc(100%-130px)] lg:w-[calc(100%-150px)]">
+                  <p className="mb-2.5 text-sm font-semibold md:text-base lg:text-lg">
+                    {data.title}
+                  </p>
+                  <p className="mb-2.5 line-clamp-2 text-xs md:text-sm lg:text-base">
+                    {data.content
+                      ? previewMarkdown(data.content).slice(0, 100)
+                      : ''}
+                  </p>
+                </div>
+                {data.image && (
+                  <div className="ml-auto h-[65px] w-[85px] shrink-0 overflow-hidden md:h-[75px] md:w-[105px] lg:h-[85px] lg:w-[125px]">
+                    <img
+                      src={data.image}
+                      className="h-full w-full object-cover"
+                    />
+                  </div>
+                )}
+              </div>
+              <ul className="mb-2.5 flex gap-3">
+                <li className="rounded-sm bg-[var(--color-gray1)] px-2 py-0.5 text-[10px] text-[var(--color-gray4)] md:text-xs lg:text-sm">
+                  태그명
+                </li>
+                <li className="rounded-sm bg-[var(--color-gray1)] px-2 py-0.5 text-[10px] text-[var(--color-gray4)] md:text-xs lg:text-sm">
+                  태그명
+                </li>
+              </ul>
+              <div className="flex items-end">
+                <span className="text-[10px] text-[var(--color-gray3)] md:text-xs lg:text-sm">
+                  {dateFormat(data.created_at)}
+                </span>
+                <div className="ml-auto flex shrink-0 gap-3">
+                  <div className="flex items-center gap-1">
+                    <Heart className="w-3.5 md:w-4 lg:w-4.5" />
+                    <span className="text-[10px] md:text-xs lg:text-sm">5</span>
+                  </div>
+                  <div className="flex items-center gap-1">
+                    <MessageSquare className="w-3.5 md:w-4 lg:w-4.5" />
+                    <span className="text-[10px] md:text-xs lg:text-sm">5</span>
+                  </div>
+                </div>
+              </div>
+            </div>
+            <div className="flex w-full items-center border-t border-[#ccc] px-3 py-2 md:px-4 md:py-2.5">
+              <Avartar />
+            </div>
+          </div>
         </div>
-      </div>
+      )}
     </>
   );
 }
